fix(build) : modalwrapper의 documnet.body -> useEffect 사용 & 빌드 실패 해결
