+++ src/app/mypage/layout.tsx
@@ -1,16 +1,16 @@
 export const metadata = {
   title: 'Next.js',
   description: 'Generated by Next.js',
-}
+};
 
 export default function RootLayout({
   children,
 }: {
-  children: React.ReactNode
+  children: React.ReactNode;
 }) {
   return (
     <html lang="en">
       <body>{children}</body>
     </html>
-  )
+  );
 }
