+++ src/app/write/layout.tsx
@@ -1,8 +1,6 @@
-import { Metadata } from 'next';
-
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+export const metadata = {
+  title: 'Next.js',
+  description: 'Generated by Next.js',
 };
 
 export default function RootLayout({
