+++ src/app/write/page.tsx
@@ -2,14 +2,14 @@
 import '@/styles/font.css';
 import { useEffect, useRef, useState } from 'react';
 import { useForm, FormProvider } from 'react-hook-form';
-import * as styles from '@/styles/country/board/write/page.css';
+import * as styles from '@/styles/write/page.css';
 import Header from '@/components/header/Header';
-import Progress from '@/components/country/board/write/Progress';
-import SearchBox from '@/components/country/board/write/firststep/SearchBox';
-import CountrySelect from '@/components/country/board/write/firststep/CountrySelect';
-import StepButton from '@/components/country/board/write/StepButton';
-import InfoInput from '@/components/country/board/write/secondstep/InfoInput';
-import ContentInput from '@/components/country/board/write/fourthstep/ContentInput';
+import Progress from '@/components/write/Progress';
+import SearchBox from '@/components/write/firststep/SearchBox';
+import CountrySelect from '@/components/write/firststep/CountrySelect';
+import StepButton from '@/components/write/StepButton';
+import InfoInput from '@/components/write/secondstep/InfoInput';
+import ContentInput from '@/components/write/fourthstep/ContentInput';
 import { BoardForm, boardFormDefault } from '@/types/country/board';
 import { useUserStore } from '@/store/userStore';
 import { usePathname, useRouter } from 'next/navigation';
