+++ src/components/write/secondstep/AttributeBox.tsx
