+++ src/components/write/secondstep/attribute/DrinkAttribute.tsx
@@ -1,38 +1,38 @@
 import Image from 'next/image';
 import { ChangeEvent } from 'react';
-import * as styles from '@/styles/country/board/write/secondstep/attribute-box.css';
+import * as styles from '@/styles/write/secondstep/attribute-box.css';
 import { useFormContext } from 'react-hook-form';
 import { BoardForm } from '@/types/country/board';
-import SmokeIcon from '@/components/common/icons/SmokeIcon';
+import DrinkIcon from '@/components/common/icons/DrinkIcon';
 
-const SmokeAttribute = () => {
+const DrinkAttribute = () => {
   const { watch, setValue } = useFormContext<BoardForm>();
 
-  const value = watch('preferSmoking');
+  const value = watch('preferDrink');
 
   const changeHandler = (event: ChangeEvent<HTMLInputElement>) => {
-    setValue('preferSmoking', Number(event.target.value));
+    setValue('preferDrink', Number(event.target.value));
   };
   return (
     <div className={styles.container}>
       <div className={styles.titleContainer}>
-        <SmokeIcon width={36} height={36} color={'#3688ff'} />
-        <span>흡연</span>
+        <DrinkIcon width={36} height={36} color={'#3688ff'} />
+        <span>음주</span>
       </div>
       <ul className={styles.radioContainer}>
         {explain.map((element, index) => (
-          <li key={'smoke' + element.text}>
+          <li key={'drink' + element.text}>
             <input
               className={styles.input}
               type="radio"
-              id={'smoke' + element.text}
+              id={'drink' + element.text}
               value={element.value}
               checked={element.value === value}
               onChange={changeHandler}
             />
             <label
               className={styles.radioGroup}
-              htmlFor={'smoke' + element.text}
+              htmlFor={'drink' + element.text}
             >
               <Image
                 className={styles.check({ select: element.value === value })}
@@ -51,9 +51,9 @@ const SmokeAttribute = () => {
 };
 
 const explain = [
-  { text: '비흡연자선호', value: 1 },
+  { text: '음주비선호', value: 1 },
   { text: '상관없음', value: 3 },
-  { text: '흡연자선호 ', value: 5 },
+  { text: '음주선호', value: 5 },
 ];
 
-export default SmokeAttribute;
+export default DrinkAttribute;
