+++ src/components/write/secondstep/attribute/GenderAttribute.tsx
@@ -1,6 +1,6 @@
 import Image from 'next/image';
 import { ChangeEvent } from 'react';
-import * as styles from '@/styles/country/board/write/secondstep/attribute-box.css';
+import * as styles from '@/styles/write/secondstep/attribute-box.css';
 import { useFormContext } from 'react-hook-form';
 import { BoardForm } from '@/types/country/board';
 import CoupleIcon from '@/components/common/icons/CoupleIcon';
