+++ src/components/write/secondstep/attribute/PhotoAttribute.tsx
@@ -1,38 +1,38 @@
 import Image from 'next/image';
 import { ChangeEvent } from 'react';
-import * as styles from '@/styles/country/board/write/secondstep/attribute-box.css';
+import * as styles from '@/styles/write/secondstep/attribute-box.css';
 import { useFormContext } from 'react-hook-form';
 import { BoardForm } from '@/types/country/board';
-import DrinkIcon from '@/components/common/icons/DrinkIcon';
+import PhotoIcon from '@/components/common/icons/PhotoIcon';
 
-const DrinkAttribute = () => {
+const PhotoAttribute = () => {
   const { watch, setValue } = useFormContext<BoardForm>();
 
-  const value = watch('preferDrink');
+  const value = watch('preferPhoto');
 
   const changeHandler = (event: ChangeEvent<HTMLInputElement>) => {
-    setValue('preferDrink', Number(event.target.value));
+    setValue('preferPhoto', Number(event.target.value));
   };
   return (
     <div className={styles.container}>
       <div className={styles.titleContainer}>
-        <DrinkIcon width={36} height={36} color={'#3688ff'} />
-        <span>음주</span>
+        <PhotoIcon width={36} height={36} color={'#3688ff'} />
+        <span>사진</span>
       </div>
       <ul className={styles.radioContainer}>
         {explain.map((element, index) => (
-          <li key={'drink' + element.text}>
+          <li key={'photo' + element.text}>
             <input
               className={styles.input}
               type="radio"
-              id={'drink' + element.text}
+              id={'photo' + element.text}
               value={element.value}
               checked={element.value === value}
               onChange={changeHandler}
             />
             <label
               className={styles.radioGroup}
-              htmlFor={'drink' + element.text}
+              htmlFor={'photo' + element.text}
             >
               <Image
                 className={styles.check({ select: element.value === value })}
@@ -51,9 +51,9 @@ const DrinkAttribute = () => {
 };
 
 const explain = [
-  { text: '음주비선호', value: 1 },
+  { text: '사진비선호', value: 1 },
   { text: '상관없음', value: 3 },
-  { text: '음주선호', value: 5 },
+  { text: '사진선호', value: 5 },
 ];
 
-export default DrinkAttribute;
+export default PhotoAttribute;
