+++ src/components/write/secondstep/attribute/SmokeAttribute.tsx
@@ -1,38 +1,38 @@
 import Image from 'next/image';
 import { ChangeEvent } from 'react';
-import * as styles from '@/styles/country/board/write/secondstep/attribute-box.css';
+import * as styles from '@/styles/write/secondstep/attribute-box.css';
 import { useFormContext } from 'react-hook-form';
 import { BoardForm } from '@/types/country/board';
-import PhotoIcon from '@/components/common/icons/PhotoIcon';
+import SmokeIcon from '@/components/common/icons/SmokeIcon';
 
-const PhotoAttribute = () => {
+const SmokeAttribute = () => {
   const { watch, setValue } = useFormContext<BoardForm>();
 
-  const value = watch('preferPhoto');
+  const value = watch('preferSmoking');
 
   const changeHandler = (event: ChangeEvent<HTMLInputElement>) => {
-    setValue('preferPhoto', Number(event.target.value));
+    setValue('preferSmoking', Number(event.target.value));
   };
   return (
     <div className={styles.container}>
       <div className={styles.titleContainer}>
-        <PhotoIcon width={36} height={36} color={'#3688ff'} />
-        <span>사진</span>
+        <SmokeIcon width={36} height={36} color={'#3688ff'} />
+        <span>흡연</span>
       </div>
       <ul className={styles.radioContainer}>
         {explain.map((element, index) => (
-          <li key={'photo' + element.text}>
+          <li key={'smoke' + element.text}>
             <input
               className={styles.input}
               type="radio"
-              id={'photo' + element.text}
+              id={'smoke' + element.text}
               value={element.value}
               checked={element.value === value}
               onChange={changeHandler}
             />
             <label
               className={styles.radioGroup}
-              htmlFor={'photo' + element.text}
+              htmlFor={'smoke' + element.text}
             >
               <Image
                 className={styles.check({ select: element.value === value })}
@@ -51,9 +51,9 @@ const PhotoAttribute = () => {
 };
 
 const explain = [
-  { text: '사진비선호', value: 1 },
+  { text: '비흡연자선호', value: 1 },
   { text: '상관없음', value: 3 },
-  { text: '사진선호', value: 5 },
+  { text: '흡연자선호 ', value: 5 },
 ];
 
-export default PhotoAttribute;
+export default SmokeAttribute;
