+++ src/styles/write/firststep/search-box.css.ts
