+++ src/styles/write/fourthstep/content-input.css.ts
