+++ src/styles/write/secondstep/attribute-box.css.ts
