+++ src/styles/write/secondstep/info-input.css.ts
