+++ src/styles/write/secondstep/people-count-box.css.ts
